@@ -5,8 +5,7 @@ import 'package:http/http.dart';
 import 'package:my_meteo/HttpClient.dart';
 import 'package:my_meteo/Weather.dart';
 
-class OWMClient extends HttpClient{
-
+class OWMClient extends HttpClient {
   final openWeatherMapUrl = "api.openweathermap.org";
   final String forecastPath = "/data/2.5/forecast";
   final String weatherPath = "/data/2.5/weather";
@@ -14,24 +13,18 @@ class OWMClient extends HttpClient{
 
   OWMClient(this.apiKey);
 
-  Future<Weather> getWeather(String city, {String stateCode, String countryCode}) async{
+  Future<Weather> getWeather(String city, {String stateCode = "", String countryCode = ""}) async {
     Map<String, String> header = HashMap();
-    stateCode = stateCode.isNotEmpty ? ","+stateCode : "";
-    countryCode = countryCode.isNotEmpty ? ","+ countryCode : "";
+    if (stateCode != "") stateCode = "," + stateCode;
+    if (countryCode != "") countryCode = "," + countryCode;
     header["q"] = city + stateCode + countryCode;
     header["appid"] = apiKey;
     var uri = Uri.http(openWeatherMapUrl, weatherPath, header);
-    Response response;
-    if(client != null){
-      response = await client.get(uri, headers: header);
-    }else{
-      response = await get(uri, headers: header);
-    }
-    if(response.statusCode == 200){
+    Response response = await get(uri, headers: header);
+    if (response.statusCode == 200) {
       return Weather.fromJson(jsonDecode(response.body));
-    }else{
+    } else {
       throw Exception('Failed to load weather');
     }
   }
-
 }
