@@ -12,12 +12,12 @@ class Weather{
   final int iconId;
 
 
-  Weather({this.name, this.temp, this.clouds, this.windSpeed, this.wheatherId,
-      this.weatherName, this.description, this.iconId});
+  Weather({required this.name, required this.temp, required this.clouds, required this.windSpeed, required this.wheatherId,
+      required this.weatherName, required this.description, required this.iconId});
 
   factory Weather.fromJson(Map<String, dynamic> json){
 
-    return Weather(name: json["name"], temp: Utils().kelvinToCelsius(json["main"]["temp"]), clouds: json["clouds"]["all"],
+    return Weather(name: json["name"], temp: Utils().kelvinToCelsius(json["main"]["temp"]), clouds: json["clouds"]["all"], windSpeed: json["wind"]["speed"],
         wheatherId: json["weather"]["id"], weatherName: json["weather"]["main"], description: json["weather"]["description"], iconId: json["weather"]["icon"]);
   }
 
